container/requirements.txt
# Demo Canvas ETL Container Dependencies
# =========================================

# Snowflake Connectivity
snowflake-snowpark-python==1.20.0
snowflake-connector-python[pandas]==3.12.0

# Web Framework for Service Endpoints
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# Data Processing
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# Data Generation
faker==22.6.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.1
httpx==0.27.0

# Type hints
typing-extensions==4.9.0



